+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Bugfixes:
 
   - `Router#start` now takes `options` and passes them to `Backbone.history.start`
   - `Model#paramRoot` returns the `baseModel.modelName` when used with STI
+  - `Array.toParam` and `Array.toQuery` no longer show up in array iterators
 
 ## 0.5.0 (March 17th, 2014)